ファイル整形
